==> scripts/06-install_packages.R <==
#### Preamble ####
# Purpose: Install necessary packages.
# Author: Yuanyi (Leo) Liu
# Date: 26 November 2024
# Contact: leoy.liu@mail.utoronto.ca
# License: MIT
# Pre-requisites: None
# Any other information needed? Make sure you are in the `Determinants-of-Wage-Variation-in-Canada` rproj


#### Install packages ####
install.packages("tidyverse")
install.packages("arrow")
install.packages("dplyr")
install.packages("rstanarm")
install.packages("here")
install.packages("ggplot2")
install.packages("kableExtra")
install.packages("modelsummary")
install.packages("janitor")
install.packages("knitr")
install.packages("testthat")
